this commit will add put method and full crud

+++ configInicialEcomponentes/src/main/java/com/example/configInicialEcomponentes/services/animes/animeDTO.java
@@ -1,4 +1,4 @@
 package com.example.configInicialEcomponentes.services.animes;
 
-public record animeDTO(int id, String name) {
+public record animeDTO(Long id, String name) {
 }
